chore(lLMSort/consolidate): Consolidate Tests and Problem
Both projects consolidated into one.
