+++ InterviewQuestions.sln
@@ -57,8 +57,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EggDropTrialQuestion", "Egg
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinkedListMergeSort", "LinkedListMergeSort\LinkedListMergeSort.csproj", "{C6226CB8-76F1-476F-A13E-C0F4F44490FD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinkedListMergeSortTester", "LinkedListMergeSortTester\LinkedListMergeSortTester.csproj", "{3B2C42BA-2278-4709-B32D-8832B85F31DD}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UtopianTree", "UtopianTree\UtopianTree.csproj", "{BE92673E-8C06-4D11-9970-5F0D682A66C8}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FirstRepeatedWordInASentence", "FirstRepeatedWordInASentence\FirstRepeatedWordInASentence.csproj", "{74AC33E6-0130-4D4A-96CE-0F8E0E67A9A5}"
@@ -177,10 +175,6 @@ Global
 		{C6226CB8-76F1-476F-A13E-C0F4F44490FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C6226CB8-76F1-476F-A13E-C0F4F44490FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C6226CB8-76F1-476F-A13E-C0F4F44490FD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3B2C42BA-2278-4709-B32D-8832B85F31DD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3B2C42BA-2278-4709-B32D-8832B85F31DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3B2C42BA-2278-4709-B32D-8832B85F31DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3B2C42BA-2278-4709-B32D-8832B85F31DD}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BE92673E-8C06-4D11-9970-5F0D682A66C8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BE92673E-8C06-4D11-9970-5F0D682A66C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BE92673E-8C06-4D11-9970-5F0D682A66C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
