+++ LinkedListMergeSort/LinkedListMergeSort.csproj
@@ -30,6 +30,9 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="nunit.framework, Version=3.9.0.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
+      <HintPath>..\packages\NUnit.3.9.0\lib\net45\nunit.framework.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -40,13 +43,17 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ArrayMergeSorter.cs" />
-    <Compile Include="LinkedList.cs" />
-    <Compile Include="LinkedListMergeSorter.cs" />
-    <Compile Include="Node.cs" />
+    <Compile Include="Problem\ArrayMergeSorter.cs" />
+    <Compile Include="Problem\LinkedList.cs" />
+    <Compile Include="Problem\LinkedListMergeSorter.cs" />
+    <Compile Include="Problem\Node.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Tests\IntArrayMergeSortTester.cs" />
+    <Compile Include="Tests\IntArrayMergeSortTestInput.cs" />
+    <Compile Include="Tests\LinkedListTester.cs" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="packages.config" />
     <None Include="README.md" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
