+++ LinkedListMergeSort/Problem/ArrayMergeSorter.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LinkedListMergeSort
+﻿namespace LinkedListMergeSort.Problem
 {
     public class ArrayMergeSorter
     {
